+++ app/Support/DateFormatter.php
@@ -44,11 +44,11 @@ class DateFormatter
         if ($diff < 60) {
             return $diff . " seconds ago";
         } elseif ($diff < 3600) {
-            return floor($diff / 60) . " minutes ago";
+            return (int) floor($diff / 60) . " minutes ago";
         } elseif ($diff < 86400) {
-            return floor($diff / 3600) . " hours ago";
+            return (int) floor($diff / 3600) . " hours ago";
         } else {
-            return floor($diff / 86400) . " days ago";
+            return (int) floor($diff / 86400) . " days ago";
         }
     }
